FEAT: prescribe ALL-CAPS type when commit includes breaking change
BREAKING CHANGE: requires type in ALL-CAPS when breaking change, whereas previously type would always be lowercase. Tools enforcing Conventional Commits as understood prior to this change may flag or fail to correctly parse commit messages indicating breaking changes under this version of the spec.

When a commit includes a breaking change, surely this is the most important thing to say about that commit. So lead with it. Indicating the breaking-change-ness of a commit it its header rather than just in its body and footer makes it easier to skim to, notice, and parse out (from abbreviated representations of the commit log) these changes.