BREAK: feat: indicate breaking changes in header via "BREAK: " prefix
BREAKING CHANGE: adds new "BREAK: " prefix before type in header in commits including breaking API changes. Tools written to the spec before this feature may fail to parse or flag as non-compliant commit messages using this feature. Breaking change commit messages written to prior versions of this spec will not be compliant with versions of the spec with this change.

Tools can regain spec compliance by adding support for parsing the new "BREAK: " prefix.

For a commit that includes API-breaking changes, that fact is the most important thing to say about the commit, so lead with it in the commit message header. Indicating breakingness in the commit message header makes API breakages more apparent on skim in more contexts.