fix: Remove spec ambiguity between body and footer
Prior to this change, if a message had a block of text after the type and description then it was ambiguous as to whether it was a body or a footer. This change resolves the issue by specifying that a footer can only follow (and therefore requires) a body, rather than a description.

Fixes #98