fix: replace "refactor!" with "feat!"
Refactoring by definition "is the process of changing a software system in such a way that it does not alter the external behavior of the code..." So, a refactoring cannot introduce a breaking change, otherwise it is not a refactoring anymore. Using `refactor` commit type and the breaking change marker `!` is a contradiction.